Reorganize hierarchy of dispatcher classes
