+++ app/services/solidus_subscriptions/dispatcher/out_of_stock_dispatcher.rb
@@ -2,9 +2,11 @@
 
 # Handles installments that cannot be processed for lack of stock.
 module SolidusSubscriptions
-  class OutOfStockDispatcher < Dispatcher
-    def dispatch
-      installments.each(&:out_of_stock)
+  module Dispatcher
+    class OutOfStockDispatcher < Base
+      def dispatch
+        installments.each(&:out_of_stock)
+      end
     end
   end
 end
